\name{expandKEGGNode}
\alias{expandKEGGNode}
\title{ Expand KEGG node of paralogues}
\description{
  The function expands KEGG node of paralogues, and is mainly used
  internally. The end-users are not expected to call it unless they know
  exactly what they are doing.
}
\usage{
expandKEGGNode(node)
}
\arguments{
  \item{node}{ An object of \code{\link{KEGGNode-class}} }
}
\author{ Jitao David Zhang \url{mailto:jitao_david.zhang@roche.com} }
