% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_graph.R
\name{extract_entry_positions}
\alias{extract_entry_positions}
\title{This function extracts the (x, y) positions of elements (entries) from a KEGG KGML XML document.}
\usage{
extract_entry_positions(xml_doc)
}
\arguments{
\item{xml_doc}{An XML document object as returned by load_kegg_kxml_file().}
}
\value{
A data.frame with columns: id, name, x, y.
}
\description{
This function extracts the (x, y) positions of elements (entries) from a KEGG KGML XML document.
}
