\name{getValue-methods}
\docType{methods}
\alias{getValue}
\alias{getValue-methods}
\alias{getValue,KEGGEdgeSubtype-method}
\title{ Get 'value' attribute}
\description{
  Get 'value' attribute, mainly used internally and is not expected to
  be called by users.
}
\section{Methods}{
\describe{
\item{object = "KEGGEdgeSubType"}{ An object of \code{\link{KEGGEdgeSubType-class}} }
}}
\keyword{methods}

