\name{KEGGPATHID2NAME}
\alias{KEGGPATHID2NAME}
\docType{data}
\title{KEGG pathway identifier to name}
\description{
  The data provides a translation mechanism between KEGG pathway identifiers,
  for instance \code{hsa04010}, and pathway names, for instance \code{MAPK
  signaling pathway}.
}
\format{
An \code{AnnDbBiMap}
}
\keyword{datasets}
