% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_graph.R
\name{kgml_to_igraph}
\alias{kgml_to_igraph}
\title{Convert KEGG KGML XML to igraph object}
\usage{
kgml_to_igraph(xml_doc)
}
\arguments{
\item{xml_doc}{An XML document object as returned by load_kegg_kxml_file().}
}
\value{
An igraph object representing the pathway graph.
}
\description{
This function parses a KEGG KGML XML document and converts it to an igraph object.
}
