% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_graph.R
\name{load_kegg_kxml_file}
\alias{load_kegg_kxml_file}
\title{Load a KEGG pathway KXML file}
\usage{
load_kegg_kxml_file(file_path)
}
\arguments{
\item{file_path}{Character. Path to the KXML file to load.}
}
\value{
An XML document object.
}
\description{
This function loads a KEGG pathway KXML file from disk and returns its contents as an XML document.
}
