\name{parseSubType}
\alias{parseSubType}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Parse KGML relation subtype }
\description{
  The function parses KGML relation subtype, called internally and not
  intended to be used by end users.
}
\usage{
parseSubType(subtype)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{subtype}{ KGML subtype node}
}
\value{
  An object of \code{\link{KEGGEdgeSubType-class}}
}
\author{ Jitao David Zhang \url{mailto:jitao_david.zhang@roche.com} }
