% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_graph.R
\name{plot_igraph_interactive}
\alias{plot_igraph_interactive}
\title{Plot igraph object interactively using plotly}
\usage{
plot_igraph_interactive(g, de_results = NULL, color_column = "pvalue")
}
\arguments{
\item{g}{An igraph object.}
}
\value{
A plotly object.
}
\description{
This function creates an interactive visualization of an igraph object using plotly.
}
