\name{top}
\alias{top}
\docType{data}
\title{
Colorectal cancer dataset
}
\description{
  Colorectal cancer dataset provided by \code{SPIA} package. It is just a
  copy during the development of \code{SPIA} package in case the package
  is not available. It will be removed when the \code{SPIA} package is
  stable.

  see the description of \code{SPIA} package.
}
\usage{data(colorectalcancerSPIA)}
\format{
    see the format of \code{SPIA} package.
}
\source{
     Yi Hong and  Kok Sun Ho and  Kong Weng Eu and  Peh Yean Cheah, A
     susceptibility gene set for early onset colorectal cancer that
     integrates diverse signaling pathways: implication for
     tumorigenesis, Clin Cancer Res, 2007, 13(4),1107-14.
}
\keyword{datasets}
